Add files via upload
Mycology is a branch of biology that deals with the study of fungi, particularly their genetic and biochemical structures, as well as their relevance to human use. Scientists in various places have also noted the medicinal properties of fungi, but not all types of fungi are beneficial - some can be quite dangerous.

We will use a decision tree to examine a dataset about fungi, extracted from the Audubon Society's Field Guide to North American Mushrooms (1981). The dataset provides detailed information about various types of mushrooms, including various physical characteristics such as cap size and stem length, along with their toxicity or edibility classification.